style: 쿼리문과 커맨드문 컨트롤러 분리
